[ng] news: update w.r.t. recent changes in testsuite harnesses (TESTS based)
* NG-NEWS (Parallel testsuite harness): The 'test-driver' auxiliary script
is always distributed when the TESTS variable is seen, even if it is not
strictly required.  Conditional and dynamic (even computed at make runtime)
content of TEST_EXTENSIONS is supported.  We are more liberal in accepting
TEST_EXTENSIONS entries (it's enough that they start with a dot), and if
any of them is invalid, an error is given at make runtime, not at automake
runtime.  Adjust some existing text to be more clear, thanks to a better
use of separating vertical and horizontal space.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>