fixup: interaction between verbatim lines and line continuation
When line continuations were involved, our hack to pass lines verbatim
to the output Makefile didn't work as expected; for example, the input

    !$(call foo,=, \
    !   long continued line still in the call)

produced in the generated Makefile an output like

    $(call foo,=, \
    !   long continued line still in the call)

rather than as the expected

    $(call foo,=, \
       long continued line still in the call)

That bug severely limited the usefulness of our hack.  Luckily, it's
pretty easy to fix.

* automake.in (file_contents_internal): Handling of !-escaped
lines moved ...
(make_paragraphs): ... here, and adjusted to cope with line
continuations.
* lib/am/parallel-tests.am: Break overly long !-escaped lines,
now that we can.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>