[ng] tests: fix spurious failure under "make distcheck"
* t/verbatim.sh: This test copied in files from the directory of Automake
internal '.am' fragments, and then overwrote one of them with 'cat >'.
But that is bound to fail under a "make distcheck", because in that case
all of the source directory (in particular the '.am' files in it) are made
read only.  This can be fixed by unlinking the file we want to overwrite
before actually writing to it.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>