cleanup: get rid of 'Automake::Configure_ac' module
Now that we unconditionally assume the Autoconf input is named
"configure.ac", it's easier and clearer to inline the checks on
the existence of such a file directly in the automake and
aclocal scripts.

This change is basically a backport of the 'v1.12-327-gee6d72a'
commit from Automake-NG.

* lib/Automake/Variable.pm (Automake::Configure_ac): Drop this import.
($configure_ac): Define to "configure.ac", unconditionally.
* automake.in: Likewise.  Also, in main code, check the existence
of 'configure.ac' directly, instead of calling the now-removed
'require_configure_ac' function.
* aclocal.in: Likewise.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>