[ng] compile: rename make variable DEFAULT_INCLUDES -> AM_DEFAULT_INCLUDES
The automake-generated variable 'DEFAULT_INCLUDES' has never been
documented, and it violates user namespace.  So rename it to
'AM_DEFAULT_INCLUDES', to make it clear that it is owned by Automake.
We prefer that name to something that would sound more "private",
like 'am__default_includes', because there are possibly legitimate
usages of that variable on the user's part.

* automake.in, lib/am/compile.am: Do the rename.
* NG-NEWS, doc/automake-ng.texi: Update.
* t/confh4.sh, t/no-extra-makefile-code.sh: Adjust.
* syntax-checks.mk (old_vars_checks): Add new check
'sc_no_DEFAULT_INCLUDES'.
(modern_DEFAULT_INCLUDES): Define to 'DEFAULT_INCLUDES'.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>