general: assume dot-starting file names are supported
Now that we've dropped support for DJGPP, Windows 95/98/ME
and MS-DOS, that should always the case.

This change is basically a backport of Automake-NG commit
'v1.12-342-ge6af355'.

* lib/am/texibuild.am: Just assume we can have directories whose
name starts with a dot, i.e., that '${am__leading_dot}' expands
to '.'.
* automake.in (require_build_directory): Likewise.
* m4/depend.m4 (AM_SET_DEPDIR): Likewise.  Now $(DEPDIR) will
be set unconditionally to '.deps'.
* m4/init.m4 (AM_INIT_AUTOMAKE): Don't AC_REQUIRE the m4 macro
'AM_SET_LEADING_DOT' anymore.
* m4/lead-dot.m4: Delete.
* Makefile.am (dist_automake_ac_DATA): Remove it.
* t/subobj11a.sh: Just assume $(DEPDIR) expands to '.deps'.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>