[ng] refactor: new make variables am__all_libs and am__all_ltlibs
This change is only a preparatory refactoring in view of future changes.

* t/all-progs.sh: Renamed ...
* t/all-prog-libs.sh: ... like this, and extended.
* automake.in (generate_makefile): Define the make vars 'am__all_libs'
and 'am__all_ltlibs' respectively from the sorted keys of the hashes
'%known_libraries' and '%known_ltlibraries'.  For consistency, and for
offering a better stability of the output Makefiles, sort the keys of
the '%known_programs' hash before using them to define the make variable
'am__all_progs'.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>